customer-service/requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.2.0
django-cors-headers==4.3.1
PyJWT==2.8.0
requests==2.31.0
